[NBS-16_create_a_gravity_module] Added UAV body and started System Test Case
Added UAV body in preperation for developing a system test case which will be used for testing new features that get
added to the simulator. It is slightly different to unit test as this is meant to test things in a baselin environment
while the unit test is meant to test things in an echo chamber.