[NBS-16_create_a_gravity_module] [NBS-16_create_a_gravity_module] Fixed System Test Simulation
Found an error in header file for Grvaity_findGravity() function. Further, there was
an issue with the GConversion_string2uint8() function so changed that to make it more robust.
Also found that the module name part of the function name had a typo for GConversions. After
all these issues where resolved, the simulation was able to run with no sig faults and exit normally.