[NBS-16_create_a_gravity_module] Debuged System Test to get changes working
Had to fix some formatting and typo errors.

The big change was how pointers work in C. The rigid body array needs to be an array of pointers
to the rigid bodies of the different types of bodies. This way, when you make a change to the
rigidBody array, the variables in the original struct are also changed. I set it up in a way
to avoid having to use any alloc's. This way things should have no risk of leaks.

Managed to get some data that looks reasonable. Still need to sanity check the DualBody gravity
model, as well as complete the verification of the nBody gravity model.