[NBS-38_create_module_for_simulating_magnetic_field] Initial set up of parameters and init function
Redeveloped how parameters are managed in the RunTest.sh script to be better
suited for the simulation setup. Further, added some additional function to
GParser and GConversions so that can load more data types.

Intention is to now test that the parameters are loaded corrcetly before doing
any more development work. This way any issues are caught early and easier to
disect.