[NBS-38_create_module_for_simulating_magnetic_field] Debugged init function.
There where minor errors but the big error was how the GParser module didn't have a
protection against writing beyond the buffer size. So what was happening was the
valueBuffer was being overwritten because the keyBuffer was full. Hence a check
was added so the user is aware of this error now and it doesn't go unoticed.

Also fixed a bug in RunTest.sh script so that it doesnr't use -d flag when creating
the TestRun folder name if -d is parsed last.